requirements.txt
langchain-core
langgraph
streamlit
python-dotenv
langchain[google-genai]
